% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%||\%}
\alias{\%||\%}
\title{Double-pipe}
\usage{
x \%||\% y
}
\arguments{
\item{x}{An object of interest, that might be \code{NULL}}

\item{y}{A default value, in case x is \code{NULL}}
}
\description{
Return default value when a value is NULL
}
\examples{
sometimes_null <- function(t) if (t == "woops") NULL else t
sometimes_null("hi there") \%||\% "bad value entered"
sometimes_null("woops") \%||\% "bad value entered"
}

