% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%+\%}
\alias{\%+\%}
\title{String concatenation}
\usage{
x \%+\% y
}
\arguments{
\item{x, y}{character vectors to be pasted}
}
\description{
Add together strings
}
\examples{
name <- "Tarak"
"Hello " \%+\% name
}

